Adds homepage backend API & homepage hero section
